Int Arrays To Doubles

+++ Gold_v1/app/src/main/java/com/glucose/arjunwatane/gold_v1/Derive.java
@@ -2,17 +2,17 @@ package com.glucose.arjunwatane.gold_v1;
 
 public class Derive {
     static final int rows=100, cols=500;
-    int[][] data;
+    double[][] data;
 
     //test constructor
     Derive(){
         //example raw wavelength
-        int[][] wave1 = new int[rows][cols];
+        double[][] wave1 = new double[rows][cols];
 
         for(int i=0; i<wave1.length; i++){
             //populate
             for(int j=0; j<wave1[0].length; j++){
-                wave1[i][j] = j+i+2;//multiply an index j by its exponent
+                wave1[i][j] = (double) j+i+2;//multiply an index j by its exponent
                 System.out.print(wave1[i][j] +" ");
             }
             System.out.println();
@@ -23,23 +23,23 @@ public class Derive {
     }
 
     //process the spectrum with first derivative
-    Derive(int[][] spectrum){
+    Derive(double[][] spectrum){
         spectrum = derivative(spectrum);
         data = spectrum;
     }
 
     //take first derivative
-    int[][] derivative(int[][] spectrum){
+    double[][] derivative(double[][] spectrum){
         for(int i=0; i<spectrum.length; i++){
             for(int j=0; j<spectrum[0].length; j++){
-                spectrum[i][j] *= j+1;
+                spectrum[i][j] *= (double) j+1;
             }
         }
         return spectrum;//return derived spectrum
     }
 
     //return derived matrix
-    public int[][] getData(){
+    public double[][] getData(){
         return data;
     }
 }
